Separate input and fix placeholders
